+++ hitable.h
@@ -6,7 +6,7 @@
 class material;
 
 struct hit_record {
-	float t;
+	double t;
 	vec3 p;
 	vec3 normal;
 	material *mat_ptr;
@@ -19,7 +19,7 @@ struct hit_record {
 
 class hitable {
 	public:
-		virtual bool hit(const ray& r, float t_min, float t_max,
+		virtual bool hit(const ray& r, double t_min, double t_max,
 				hit_record& rec) const = 0;
 };
